+++ lib/atom_notifier.dart
@@ -1,3 +1,4 @@
 library atom_notifier;
 
 export "./notifier.dart";
+export "./application_states.dart";