+++ CHANGELOG.md
@@ -1,3 +1,3 @@
-## 0.0.1
-
-* TODO: Describe initial release.
+## 1.0.0
+- Added default reactive components
+- Added default application states